ref(profiling): restructure profileDetails component
