+++ static/app/views/profiling/profileDetails/index.tsx
@@ -1,0 +1,1 @@
+export {default as default} from './profileDetails';
